requirements.txt
asyncpg==0.29.0
psycopg2==2.9.9 
pydantic==2.9.1
sqlalchemy==2.0.34 
python-decouple==3.8